Use the compatible constructor Buffer(byte[], int, int) when Buffer() does not exist
