+++ gradle/dependency-locks/compileClasspath.lockfile
@@ -1,4 +1,6 @@
 # This is a Gradle generated file for dependency locking.
 # Manual edits can break the build and are not advised.
 # This file is expected to be part of source control.
-org.embulk:embulk-api:0.10.8
+org.embulk:embulk-api:0.10.28
+org.msgpack:msgpack-core:0.8.11
+org.slf4j:slf4j-api:1.7.30
