+++ src/main/java/org/embulk/util/file/EmptyBuffer.java
@@ -18,67 +18,40 @@ package org.embulk.util.file;
 
 import org.embulk.spi.Buffer;
 
-final class EmptyBuffer extends Buffer {
-    EmptyBuffer() {
+final class EmptyBuffer {
+    private EmptyBuffer() {}
+
+    static Buffer getInstance() {
+        return Holder.INSTANCE;
     }
 
-    @SuppressWarnings("deprecation")
-    @Override
-    public byte[] array() {
-        return EMPTY_BYTES;
+    private static class Holder {  // Initialization-on-demand holder idiom.
+        private static final Buffer INSTANCE = createInstance();
+
+        private static Buffer createInstance() {
+            // EmptyBufferCompat and EmptyBufferUpToDate are implemented as separate independent classes while
+            // they have the same methods so that this library can work both with Embulk v0.9 and v0.10.
+            //
+            // EmptyBufferUpToDate depends on the constructor |Buffer()| of the latest abstract org.embulk.spi.Buffer.
+            // https://github.com/embulk/embulk/blob/v0.10.27/embulk-api/src/main/java/org/embulk/spi/Buffer.java#L38-L39
+            //
+            // It is basically recommended in Embulk v0.10 and later, but it did not exist in Embulk v0.9 or earlier.
+            // Embulk v0.9 had only the constructor |Buffer(byte[], int, int)| of org.embulk.spi.Buffer.
+            // https://github.com/embulk/embulk/blob/v0.9.23/embulk-core/src/main/java/org/embulk/spi/Buffer.java
+            //
+            // EmptyBufferCompat is implemented to depend on |Buffer(byte[], int, int)| to work with Embulk v0.9.
+            //
+            // They are implemented as separate independent classes because one can fail when the class is initialized
+            // if its constructor is depending on a non-existing super-class constructor. They would never fail unless
+            // the class is initialized / loaded.
+            try {
+                Buffer.class.getConstructor();  // throws NoSuchMethodException if working with v0.9.
+                return new EmptyBufferUpToDate();
+            } catch (final NoSuchMethodException ex) {
+                // Pass-through.
+            }
+
+            return new EmptyBufferCompat();
+        }
     }
-
-    @Override
-    public int offset() {
-        return 0;
-    }
-
-    @Override
-    public Buffer offset(final int offset) {
-        return this;
-    }
-
-    @Override
-    public int limit() {
-        return 0;
-    }
-
-    @Override
-    public Buffer limit(final int limit) {
-        return this;
-    }
-
-    @Override
-    public int capacity() {
-        return 0;
-    }
-
-    @Override
-    public void setBytes(int index, byte[] source, int sourceIndex, int length) {
-        return;
-    }
-
-    @Override
-    public void setBytes(int index, Buffer source, int sourceIndex, int length) {
-        return;
-    }
-
-    @Override
-    public void getBytes(int index, byte[] dest, int destIndex, int length) {
-        return;
-    }
-
-    @Override
-    public void getBytes(int index, Buffer dest, int destIndex, int length) {
-        return;
-    }
-
-    @Override
-    public void release() {
-        return;
-    }
-
-    static final EmptyBuffer INSTANCE = new EmptyBuffer();
-
-    private static final byte[] EMPTY_BYTES = new byte[0];
 }
